requirements.txt
streamlit
google-genai
scikit-learn==1.3.2
joblib==1.3.2
numpy<2
pandas==2.1.4
datasets==2.14.5
pyarrow==12.0.1
